Refactor views tests to match views module
