+++ app/views/info.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, Response
 
-blueprint = Blueprint('healthcheck', __name__, template_folder='templates')
+blueprint = Blueprint('info', __name__, template_folder='templates')
 
 
 @blueprint.route('/info', methods=["GET"])
