+++ tests/views/test_info.py
@@ -1,0 +1,3 @@
+def test_get_healthcheck(client):
+    response = client.get('/info')
+    assert response.status_code == 200
