+++ tests/views/test_load_sample.py
